refactor: restructure project into modular packages
- Split monolithic main.go into specialized packages

- Create internal/handlers package for HTTP handlers

- Create internal/util package for shared utilities

- Move API config and types to appropriate packages

- Standardize logging across packages

- Move JSON response handling to utility package

- Update import dependencies

- Clean up main.go to focus on server setup
