+++ go.mod
@@ -3,7 +3,7 @@ module github.com/kwekkwekpatu/chirpy
 go 1.22.5
 
 require (
-	github.com/google/uuid v1.6.0 // indirect
-	github.com/joho/godotenv v1.5.1 // indirect
-	github.com/lib/pq v1.10.9 // indirect
+	github.com/google/uuid v1.6.0
+	github.com/joho/godotenv v1.5.1
+	github.com/lib/pq v1.10.9
 )
